+++ CHANGELOG.md
@@ -184,6 +184,7 @@ Modifications can only be done via builders.
 - Optimize produce performance
 - Allow to disable extra checks
 - Optimize nomination - Certain nodes should produce directly and should not run through the nomination and production process
+- Allow to disable on.change and on.changeRecursive
 
 ### Fixed